+++ common/src/main/java/com/xtracr/realcamera/mixins/MixinItem.java
@@ -2,30 +2,27 @@ package com.xtracr.realcamera.mixins;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import com.xtracr.realcamera.camera.CameraController;
+import com.xtracr.realcamera.utils.RaycastUtils;
 
-import net.minecraft.client.MinecraftClient;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
-import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.hit.BlockHitResult;
+import net.minecraft.world.RaycastContext;
+import net.minecraft.world.World;
 
 @Mixin(Item.class)
 public abstract class MixinItem {
     
-    @ModifyVariable(method = "raycast", at = @At("STORE"), ordinal = 0)
-    private static Vec3d getCameraPosition(Vec3d vec3d) {
-        return (CameraController.isActive() ? vec3d.add(CameraController.getCameraOffset()) : vec3d);
+    @Inject(method = "raycast", at = @At("HEAD"), cancellable = true)
+    private static void coverRaycast(World world, PlayerEntity player, RaycastContext.FluidHandling fluidHandling, CallbackInfoReturnable<BlockHitResult> cInfo){
+        if (CameraController.isActive()) {
+            RaycastUtils.update(player, 25.0D, 1.0F);
+            cInfo.setReturnValue(world.raycast(RaycastUtils.getRaycastContext(RaycastContext.ShapeType.OUTLINE, fluidHandling, player)));
+        }
     }
-    
-    @ModifyVariable(method = "raycast", at = @At("STORE"), ordinal = 0)
-    private static float getCameraPitch(float f) {
-        return (CameraController.doCrosshairRotate() ? MinecraftClient.getInstance().getEntityRenderDispatcher().camera.getPitch() : f);
-    }
-    
-    @ModifyVariable(method = "raycast", at = @At("STORE"), ordinal = 1)
-    private static float getCameraYaw(float g) {
-        return (CameraController.doCrosshairRotate() ? MinecraftClient.getInstance().getEntityRenderDispatcher().camera.getYaw() : g);
-    }
-    
+
 }