+++ common/src/main/java/com/xtracr/realcamera/utils/MathUtils.java
@@ -17,4 +17,9 @@ public class MathUtils {
         return new Vec3d(xRot, yRot, zRot);
     }
 
+    public static Vec3d getIntersectionPoint(Vec3d planePoint, Vec3d planeNormal, Vec3d linePoint, Vec3d lineNormal) {
+        double distance = planeNormal.dotProduct(planePoint.subtract(linePoint)) / planeNormal.dotProduct(lineNormal);
+		return linePoint.add(lineNormal.multiply(distance));
+    }
+
 }